Fix integer undeflow

+++ sources/EatHook.cpp
@@ -36,7 +36,7 @@ bool PLH::EatHook::hook() {
 		MemoryProtector protector(m_trampoline, 64, ProtFlag::R | ProtFlag::W | ProtFlag::X, *this, false);
 
 		PLH::ADisassembler::writeEncoding(makeAgnosticJmp(m_trampoline, m_fnCallback), *this);
-		offset = m_trampoline - m_moduleBase;
+		offset = static_cast<size_t>(m_trampoline - m_moduleBase);
 
 		Log::log("EAT hook offset is > 32bit's. Allocation of trampoline necessary", ErrorLevel::INFO);
 	}
